requirements.txt
pandas
numpy
scikit-learn
xgboost
fastapi
uvicorn
pydantic
joblib
pyyaml
matplotlib
seaborn
pytest
httpx